+++ src/lib.rs
@@ -41,8 +41,8 @@ pub use crate::algorithms::paths::{
     EulerianClassification, diameter, distances, eccentricity, eulerian_path, is_eulerian, radius,
 };
 pub use crate::algorithms::properties::{
-    assortativity_degree, avg_nearest_neighbor_degree, closeness, count_triangles, density, girth,
-    harmonic_centrality, mean_distance, reciprocity, transitivity_local_undirected,
+    assortativity_degree, avg_nearest_neighbor_degree, betweenness, closeness, count_triangles,
+    density, girth, harmonic_centrality, mean_distance, reciprocity, transitivity_local_undirected,
     transitivity_undirected,
 };
 pub use crate::algorithms::traversal::{BfsTree, bfs, bfs_tree, dfs};